change Id type to long

+++ Backend/Model/Coach.cs
@@ -6,10 +6,10 @@ namespace Backend.Model;
 public class Coach
 {
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
     public NationalityEnum Nationality { get; set; }
     public Gender Gender { get; set; }
     public Team Team { get; set; }
-    public int TeamId { get; set; }
+    public long TeamId { get; set; }
 }
